Boxing, enumerações, datas e valores monetários

+++ banco/src/com/paty/banco/modelo/pagamento/Boleto.java
@@ -2,21 +2,24 @@ package com.paty.banco.modelo.pagamento;
 
 import com.paty.banco.modelo.Pessoa;
 
+import java.math.BigDecimal;
+
 public class Boleto implements DocumentoPagavel, DocumentoEstornavel {
 
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean pago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
 
          }
 
     @Override
-    public double getValorTotal() {
-         return valor;
+    public BigDecimal getValorTotal() {
+
+        return valor;
     }
 
     @Override
@@ -34,4 +37,5 @@ public class Boleto implements DocumentoPagavel, DocumentoEstornavel {
         pago = true;
 
     }
+
 }
